# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: lunchrush delivery-api
  version: tier-6.0.0
  description: >
    Lifecycle de entrega, cadastro de entregador, atribuição e emissão de
    token do delivery-api (internal/platform/httpapi/server.go). Cobre só
    o delivery-api; tracking-ingest e tracking-projector não estão neste
    documento (endpoints de tracking, ver internal/tracking e
    cmd/tracking-projector/http.go, ficam fora de escopo deste OpenAPI por
    enquanto).
paths:
...
  /auth/tokens:
    post:
      summary: >
        emite um JWT HS256 para caller/tracking-ingest/tracking-projector
        validarem (delivery-api é a identidade raiz do sistema, ver
        docs/adr/0005-tokens-assinados-em-vez-de-oidc.md)
      parameters:
        - name: X-Admin-Secret
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [caller]
              properties:
                caller: { type: string }
      responses:
        "201":
          description: token emitido
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        "401": { description: X-Admin-Secret ausente ou inválido }
        "422": { description: caller ausente }
  /couriers:
    post:
      summary: cadastra um entregador, indisponível por padrão
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "201":
          description: entregador cadastrado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Courier" }
  /couriers/{id}/availability:
    post:
      summary: declara disponibilidade do entregador
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [available]
              properties:
                available: { type: boolean }
      responses:
        "204": { description: disponibilidade atualizada }
        "404": { description: entregador não encontrado }
components:
  schemas:
...
    Courier:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        available: { type: boolean }